requirements.txt
boto3
botocore
pulsar-cloud
pandas
uvloop